### 2019-06-24 #14 @whedon
[PRE REVIEW]: A general-purpose toolbox for efficient Kronecker-based learning
**Submitting author:** @michielstock (<a href="http://orcid.org/0000-0003-0903-6061">Michiel Stock</a>)
**Repository:** <a href="https://github.com/michielStock/Kronecker.jl" target ="_blank">https://github.com/michielStock/Kronecker.jl</a>

Thanks for submitting your paper to JuliaCon @michielstock. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@michielstock if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
