+++ paper/paper.yml
@@ -1,0 +1,23 @@
+title: "A general-purpose toolbox for efficient Kronecker-based learning"
+keywords:
+  - Pairwise learning
+  - Kronecker product
+  - Linear algebra
+authors:
+  - name: Michiel Stock
+    orcid: 0000-0003-0903-6061
+    affiliation: 1
+  - name: Tapio Pahikkala
+    affiliation: 2
+  - name: Antti Airola
+    affiliation: 2
+  - name: Bernard De Baets
+    orcid: 0000-0002-3876-620X
+    affiliation: 1
+affiliations:
+  - name: KERMIT, Department of Data Analysis and Mathematical Modelling, Ghent University, Belgium
+    index: 1
+  - name:  Department of Future Technologies
+    index: 2
+date: 23 June 2019
+bibliography: ref.bib
