### 2019-06-24 #15 @whedon
[REVIEW]: A general-purpose toolbox for efficient Kronecker-based learning
**Submitting author:** @michielstock (<a href="http://orcid.org/0000-0003-0903-6061">Michiel Stock</a>)
**Repository:** <a href="https://github.com/michielStock/Kronecker.jl" target ="_blank">https://github.com/michielStock/Kronecker.jl</a>

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
