### 2019-06-25 MichielStock/Kronecker.jl#6 @christianpeel
Improve Documentation: Statement of need, examples, functionality
The JuliaCon proceedings [review](https://github.com/JuliaCon/proceedings-review/issues/15) has several documentation-related questions for reviewers; it would help me out if you'd address the following:
> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

For example you could state that your package allows Kronecker products to be much faster.

>  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Moving an example from your notebook into the readme would satisfy this.

>  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

You could add Documenter.jl documentation, though I think you should start with adding the stuff above to README.md and adding a link to your Jupyter notebook.

I think Kronecker.jl is an excellent package!
