### 2019-06-25 MichielStock/Kronecker.jl#7 @christianpeel
Add DOIs for references
The JuliaCon proceedings [review](https://github.com/JuliaCon/proceedings-review/issues/15) asks for DOIs for each reference in your paper:
> **References**: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

It's not clear to me how critical this requirement is, yet it's easy enough to add them, for example the DOI for your first reference can be found on [IEEE Xplore](https://ieeexplore.ieee.org/abstract/document/7999226).
