### 2019-06-25 MichielStock/Kronecker.jl#6 @matbesancon
Documenter would be a nice step, but the minimum IMO is having docstrings + README well-detailed
