### 2019-06-25 MichielStock/Kronecker.jl#7 @matbesancon
it's a good practice to have DOIs whenever available, there is a whedon command available in https://github.com/JuliaCon/proceedings-review/issues/15
It's not fail-proof, but good indications already
