### 2019-06-25 fdc8c1e
:pencil2: add doi references
