+++ paper/ref.bib
@@ -1,6 +1,35 @@
+@techreport{Sch2013,
+abstract = {In this paper, we review basic properties of the Kronecker product, and give an overview of its history and applications. We then move on to introducing the symmetric Kronecker product, and we derive sev- eral of its properties. Furthermore, we show its application in finding search directions in semidefinite programming.},
+author = {Sch{\"{a}}cke, Kathrin},
+file = {:Users/michielstock/Library/Application Support/Mendeley Desktop/Downloaded/Sch{\"{a}}cke - 2013 - On the Kronecker Product.pdf:pdf},
+pages = {1--35},
+title = {{On the Kronecker Product}},
+url = {https://www.math.uwaterloo.ca/{~}hwolkowi/henry/reports/kronthesisschaecke04.pdf},
+year = {2013}
+}
+@article{Leskovec2008,
+abstract = {How can we model networks with a mathematically tractable model that allows for rigorous analysis of network properties? Networks exhibit a long list of surprising properties: heavy tails for the degree distribution; small diameters; and densification and shrinking diameters over time. Most present network models either fail to match several of the above properties, are complicated to analyze mathematically, or both. In this paper we propose a generative model for networks that is both mathematically tractable and can generate networks that have the above mentioned properties. Our main idea is to use the Kronecker product to generate graphs that we refer to as "Kronecker graphs". First, we prove that Kronecker graphs naturally obey common network properties. We also provide empirical evidence showing that Kronecker graphs can effectively model the structure of real networks. We then present KronFit, a fast and scalable algorithm for fitting the Kronecker graph generation model to large real networks. A naive approach to fitting would take super- exponential time. In contrast, KronFit takes linear time, by exploiting the structure of Kronecker matrix multiplication and by using statistical simulation techniques. Experiments on large real and synthetic networks show that KronFit finds accurate parameters that indeed very well mimic the properties of target networks. Once fitted, the model parameters can be used to gain insights about the network structure, and the resulting synthetic graphs can be used for null- models, anonymization, extrapolations, and graph summarization.},
+archivePrefix = {arXiv},
+arxivId = {0812.4905},
+author = {Leskovec, Jure and Chakrabarti, Deepayan and Kleinberg, Jon and Faloutsos, Christos and Ghahramani, Zoubin},
+eprint = {0812.4905},
+file = {:Users/michielstock/Library/Application Support/Mendeley Desktop/Downloaded/Leskovec et al. - 2008 - Kronecker graphs an approach to modeling networks.pdf:pdf},
+isbn = {1532-4435},
+issn = {0012365X},
+journal = {Journal of Machine Learning Research},
+keywords = {Kronecker graphs,graph genera-tors,graph mining,network analysis,network evolution,network models,social networks},
+pages = {985--1042},
+pmid = {17746742},
+title = {{Kronecker graphs: an approach to modeling networks}},
+url = {https://cs.stanford.edu/{~}jure/pubs/kronecker-jmlr10.pdf http://arxiv.org/abs/0812.4905},
+volume = {11},
+year = {2008}
+}
 @article{Stock2017tskrr,
 author = {Stock, Michiel and Pahikkala, Tapio and Airola, Antti and {De Baets}, Bernard and Waegeman, Willem},
+doi = {10.1162/neco_a_01096},
 journal = {Neural Computation},
+number = {8},
 pages = {2245--2283},
 title = {{A comparative study of pairwise learning methods based on kernel ridge regression}},
 volume = {30},
@@ -11,6 +40,7 @@ abstract = {Kronecker product kernel provides the standard approach in the kerne
 archivePrefix = {arXiv},
 arxivId = {1601.01507},
[…]
