### 2019-06-25 #15 @MichielStock
Are added to the `ref.bib`, but these don't appear in the paper, as dictated by the style file.
