### 2019-06-25 7b3ecb9
:page_with_curl: documentating the readme
