### 2019-06-25 fd82942
:wrench: improve `Base.show`
