### 2019-06-26 MichielStock/Kronecker.jl#6 @MichielStock
I have already updated the `readme.md` with an example and link to the notebook.

A real-life application will be developed for the juliacon presentation.

Would like write complete documentation when the package is a bit more stable, but I will work to improve the docstrings.
