### 2019-06-26 MichielStock/Kronecker.jl#7 @christianpeel
Thanks!
