### 2019-06-26 MichielStock/Kronecker.jl#6 @christianpeel
It's a bit fuzzy to me how much documentation the proceedings want; what you have looks good to me.

Thanks
