### 2019-06-26 #15 @christianpeel
I recommend that the paper be **accepted as-is**.
