### 2019-06-29 #15 @mcognetta
**Recommendation:** accept with **minor modifications**.

-----------------------------------------------------------------

Minor paper comments:
1. 'vec trick' is two words in one place but one word (hyphenated) in another.
2. In Section 3:
> To fully make use the power of Julia, we will explore three directions.

Should be '...make use _of_ the...'

3. The citation for vec trick should be given at the first mention of vec trick (but including it for both is also fine).

Minor code comments:
1. The `show` method should be improved to be similar to printing large dense matrices. Currently it does nothing.
2. To convert to a dense matrix, `collect` must be called. It makes sense to have a `Matrix` constructor as well.
3. Many instances of unnecessary `where` statements when the type parameter is unused. For example:
https://github.com/MichielStock/Kronecker.jl/blob/a1d9d068555ff74cf99771bdbc68b71e8a90a7be/src/base.jl#L54
https://github.com/MichielStock/Kronecker.jl/blob/a1d9d068555ff74cf99771bdbc68b71e8a90a7be/src/base.jl#L77
