### 2019-06-30 11006f7
:pencil2: textual remakrs juliacon paper
