+++ src/base.jl
@@ -1,8 +1,7 @@
-abstract type GeneralizedKroneckerProduct
+abstract type GeneralizedKroneckerProduct <: AbstractMatrix{Real}
 end
 
-abstract type KroneckerProduct <: GeneralizedKroneckerProduct
-end
+Base.IndexStyle(::Type{<:GeneralizedKroneckerProduct}) = IndexLinear()
 
 # TODO: make diagonal Kronecker system
 # TODO: document functions
@@ -11,39 +10,61 @@ end
 
 # QUESTION: allow for different types of matrices?
 
-struct KroneckerProductArray{T} <: KroneckerProduct where T <: AbstractArray
-    A::T
-    B::T
+# general Kronecker product between two matrices
+struct KroneckerProduct <: GeneralizedKroneckerProduct
+    A::AbstractMatrix
+    B::AbstractMatrix
 end
 
 """
-    issquare(A::Array{T,2}) where T <: Real
+    issquare(A::AbstractMatrix) where T <: Real
 
 Checks if an array is a square matrix.
 """
-function issquare(A::Array{T,2}) where T <: Real
+function issquare(A::AbstractMatrix)
     m, n = size(A)
     return m == n
 end
 
+# general Kronecker product between two matrices
+struct SquareKroneckerProduct <: GeneralizedKroneckerProduct
+    A::AbstractMatrix
+    B::AbstractMatrix
+    function SquareKroneckerProduct(A, B)
+        if issquare(A) && issquare(B)
+            new(A, B)
+        else
+            throw(DimensionMismatch("SquareKroneckerProduct is only for when all matrices are square"))
+        end
+    end
+end
+
+issquare(K::SquareKroneckerProduct) = true
+
+KronProd = Union{KroneckerProduct, SquareKroneckerProduct}
+
 """
-    kronecker(A::T, B::T) where T <: AbstractArray
+    kronecker(A::AbstractMatrix, B::AbstractMatrix)
 
 Construct a Kronecker product object between two arrays. Does not evaluate the
 Kronecker product explictly!
 """
-function kronecker(A::T, B::T) where T <: AbstractArray
-    return KroneckerProductArray(A, B)
+function kronecker(A::AbstractMatrix, B::AbstractMatrix)
+    if issquare(A) && issquare(B)
+        return SquareKroneckerProduct(A, B)
+    else
+        return KroneckerProduct(A, B)
+    end
 end
 
 """
-    ⊗(A::T, B::T) where T <: AbstractArray
+    ⊗(A::AbstractMatrix, B::AbstractMatrix)
 
 Construct a Kronecker product object between two arrays. Does not evaluate the
 Kronecker product explictly!
 """
-function ⊗(A::T, B::T) where T <: AbstractArray
-    return KroneckerProductArray(A, B)
+function ⊗(A::AbstractMatrix, B::AbstractMatrix)
+    return kronecker(A, B)
 end
 
 """
@@ -51,74 +72,74 @@ end
 
 Obtain the two matrices of a `KroneckerPoduct` object.
 """
-function getmatrices(K::T) where T <: KroneckerProduct
+function getmatrices(K::T) where T <: KronProd
     A = K.A
     B = K.B
     return A, B
 end
 
 """
-     Base.:size(K::T) where T <: KroneckerProductArray
+     Base.:size(K::T) where T <: KroneckerProduct
 
 Get the size of a `KroneckerPoduct` object.
 """
-function Base.:size(K::T) where T <: KroneckerProductArray
+function Base.:size(K::T) where T <: KronProd
     A, B = getmatrices(K)
[…]
