### 2019-06-30 2935dab
:wrench: update indexed kronecker products
