### 2019-06-30 f471c36
:wrench: update shifted Kronecker
