+++ src/Kronecker.jl
@@ -1,7 +1,8 @@
 module Kronecker
 
-export KroneckerProduct, KroneckerProductArray, EigenKroneckerProduct, ShiftedKroneckerProduct
-export issquare, getmatrices, size, getindices, order
+# TODO types!
+export GeneralizedKroneckerProduct, AbstractKroneckerProduct, SquareKroneckerProduct, EigenKroneckerProduct, ShiftedKroneckerProduct
+export issquare, getmatrices, size, getindices, order, issymmetric
 export ⊗, kronecker
 export tr, det, collect, inv, *, mult!, eigen, \, /, adjoint
 export getindex
