### 2019-06-30 #15 @MichielStock
Thank you for all the feedback and suggestions. I updated the code today, providing better integration with the `AbstractArray` interface. I will update the documentation and push to master branch.
