### 2019-07-01 080904e
:wrench: fix issue shifted Kronecker with non-symmetric matrices
