+++ src/Kronecker.jl
@@ -3,8 +3,8 @@ module Kronecker
 # TODO types!
 export GeneralizedKroneckerProduct, AbstractKroneckerProduct, SquareKroneckerProduct, EigenKroneckerProduct, ShiftedKroneckerProduct
 export issquare, getmatrices, size, getindices, order, issymmetric
-export ⊗, kronecker
-export tr, det, collect, inv, *, mult!, eigen, \, /, adjoint
+export ⊗, kronecker, Matrix
+export tr, det, collect, inv, *, mult!, eigen, \, /, adjoint, solve
 export getindex
 
 export genvectrick, genvectrick!
