### 2019-07-04 6e5d909
:tada: Kronecker powers!
