### 2019-07-04 4f0f856
:art: logo + update `read.me`
