### 2019-07-09 fb1b5fe
:clock3: update benchmark notebook
