### 2019-07-09 #15 @MichielStock
I have done the changes and merged them into the `master` branch.
