### 2019-07-14 #15 @christianpeel
I very much like the plot of benchmark times that you added to the Readme.
