### 2019-07-23 72306df
Add elementwise conjugation of kronecker products
exported conj
