+++ test/runtests.jl
@@ -23,6 +23,7 @@ X = kron(A, B)  # true result
 @test tr(kronprod) ≈ tr(X)
 @test det(kronprod) ≈ det(X)
 @test collect(transpose(kronprod)) ≈ transpose(X)
+@test collect(conj(kronprod)) ≈ conj(X)
 @test collect(kronprod') ≈ X'
 @test collect(inv(kronprod)) ≈ inv(X)
 @test all(kronprod * v .≈ X * v)
