+++ src/indexedkroncker.jl
@@ -1,4 +1,3 @@
-#Index = Union{UnitRange{I}, Array{I,1}} where I <: Int
 Index = Union{UnitRange{I}, AbstractVector{I}} where I <: Int #TODO: make for general indices
 
 struct IndexedKroneckerProduct <: GeneralizedKroneckerProduct
@@ -7,47 +6,56 @@ struct IndexedKroneckerProduct <: GeneralizedKroneckerProduct
     q::Index
     r::Index
     t::Index
-    function IndexedKroneckerProduct(K, p, q, r, t)
-        order(K) == 2 || throw(DimensionMismatch("Indexed Kronecker only implemented for second order Kronecker systems"))
-        length(p) == length(q) && length(r) == length(t) || throw(DimensionMismatch("Indices should have matching lengths"))
-        minimum((minimum.((p, q, r, t)))) > 0 || throw(BoundsError("Negative indices not allowed"))
+    function IndexedKroneckerProduct(K, p::Index, q::Index, r::Index, t::Index)
+        if order(K) != 2
+            throw(DimensionMismatch(
+                "Indexed Kronecker only implemented for second order Kronecker systems"
+            ))
+        elseif !(length(p) == length(q) && length(r) == length(t))
+            throw(DimensionMismatch("Indices should have matching lengths"))
+        elseif minimum((minimum.((p, q, r, t)))) <= 0
+            throw(BoundsError("Negative indices not allowed"))
+        end
         A, B = getmatrices(K)
         (m, n) = size(B)
         (k, l) = size(A)
-        maximum(p) ≤ m && maximum(q) ≤ k || throw(BoundsError("Indices exeed matrix bounds"))
-        maximum(r) ≤ n && maximum(t) ≤ l || throw(BoundsError("Indices exeed matrix bounds"))
-        new(K, p, q, r, t)
+        if !(maximum(p) ≤ m && maximum(q) ≤ k)
+            throw(BoundsError("Indices exeed matrix bounds"))
+        end
+        if !(maximum(r) ≤ n && maximum(t) ≤ l)
+            throw(BoundsError("Indices exeed matrix bounds"))
+        end
+        return new(K, p, q, r, t)
     end
 end
 
-function getindices(K::IndexedKroneckerProduct)
-    return K.p, K.q, K.r, K.t
-end
+# Best way to do multi-line strings?
+
+_dm(str) = throw(DimensionMismatch(str))
+_ibe(str) = throw(BoundsError(str))
+
+getindices(K::IndexedKroneckerProduct) = K.p, K.q, K.r, K.t
 
 getmatrices(K::IndexedKroneckerProduct) = getmatrices(K.K)
 
-function Base.:size(K::IndexedKroneckerProduct)
-    return (length(K.p), length(K.t))
-end
+Base.size(K::IndexedKroneckerProduct) = (length(K.p), length(K.t))
 
-function Base.:getindex(K::IndexedKroneckerProduct, i::Int, j::Int)
+function Base.getindex(K::IndexedKroneckerProduct, i::Int, j::Int)
     A, B = getmatrices(K)
     p, q, r, t = getindices(K)
     return B[p[i],r[j]] * A[q[i],t[j]]
 end
 
-function Base.:getindex(K::T, p::Index, q::Index, r::Index, t::Index) where T <: AbstractKroneckerProduct
+function Base.getindex(K::AbstractKroneckerProduct, p::Index, q::Index, r::Index, t::Index)
     N, M = getmatrices(K)
     a, b = size(M)
     c, d = size(N)
     return IndexedKroneckerProduct(K, p, q, r, t)
 end
 
-function Base.:eltype(K::T) where T <: IndexedKroneckerProduct
-    return eltype(K.K)
[…]
