### 2019-07-23 eead3c6
Update Kronecker.jl
