### 2019-07-26 1f44727
:tada: add Travis file
