+++ .travis.yml
@@ -1,0 +1,8 @@
+codecov: true
+coveralls: true
+
+language: julia
+julia:
+  - nightly
+  - 1.1
+  - 1.0.3
