### 2019-07-26 7f399d4
added composite types to KroneckerProduct and Square Kronecker product. Checking for squareness is disable atm
