### 2019-07-26 cca18ba
:tada: add `logdet`
