+++ src/Kronecker.jl
@@ -4,7 +4,7 @@ module Kronecker
 export GeneralizedKroneckerProduct, AbstractKroneckerProduct, SquareKroneckerProduct, EigenKroneckerProduct, ShiftedKroneckerProduct
 export issquare, getmatrices, size, getindices, order, issymmetric
 export ⊗, kronecker, Matrix
-export tr, det, collect, inv, *, mult!, eigen, \, /, adjoint, transpose, conj, solve
+export tr, det, logdet, collect, inv, *, mult!, eigen, \, /, adjoint, transpose, conj, solve
 export getindex
 
 export genvectrick, genvectrick!
