### 2019-07-26 9f996af
fixed checking for square matrices by adding an outer constructor
