+++ src/base.jl
@@ -31,16 +31,18 @@ end
 struct SquareKroneckerProduct{T<:AbstractMatrix, S<:AbstractMatrix} <: AbstractKroneckerProduct
     A::T
     B::S
-    #=function SquareKroneckerProduct{T, S}(A::T, B::S) where {T<:AbstractMatrix, S<:AbstractMatrix}
+    function SquareKroneckerProduct{T,S}(A::T, B::S) where {T<:AbstractMatrix, S<:AbstractMatrix}
         if issquare(A) && issquare(B)
-            return new{T,S}(A, B)
+            return new(A, B)
         else
             throw(DimensionMismatch(
                 "SquareKroneckerProduct is only for when all matrices are square",
             ))
         end
-    end=#
+    end
 end
+
+SquareKroneckerProduct(A::T, B::S) where {T<:AbstractMatrix, S<:AbstractMatrix} = SquareKroneckerProduct{T,S}(A,B);
 
 issquare(K::SquareKroneckerProduct) = true
 LinearAlgebra.:issymmetric(K::SquareKroneckerProduct) = issymmetric(K.A) && issymmetric(K.B)
