### 2019-07-26 414d449
:wrench: Add `AbstractSquareKronecker` in the node tree
