### 2019-07-26 24d3b55
:truck: reorganizing test suite
