### 2019-07-27 e015e45
Extra tests and some minor changes
