+++ test/eigen.jl
@@ -1,14 +1,9 @@
 using Random, LinearAlgebra
 
-@testset "eigen" begin
-    rng, P, Q = MersenneTwister(123456), 3, 5
-
-    # Generate some positive definite matrices so that logdet can be tested.
-    A_, B_ = randn(rng, P, P), randn(rng, Q, Q)
-    A, B = Symmetric(A_ * A_' + I), Symmetric(B_ * B_' + I)
+# Standardised tests for the eigen decomposition of a square kronecker product
+function eigen_tests(rng, C::SquareKroneckerProduct)
 
     # Check approximate correctness of decomposition
-    C = kronecker(A, B)
     λ, Γ = eigen(C)
     @test Γ * Diagonal(λ) * Γ' ≈ C
 
@@ -16,13 +11,34 @@ using Random, LinearAlgebra
     σ² = abs2(randn(rng))
     λ′, Γ′ = (eigen(C) + σ² * I)
     @test Γ′ * Diagonal(λ′) * Γ′' ≈ Matrix(C + σ² * I)
+    @test σ² * I + C == C + σ² * I
 
     # Test with decomposition
-    v = randn(rng, P * Q)
-    @test eigen(C) \ v ≈ Matrix(C) \ v
+    v = randn(rng, length(λ))
+    @test eigen(C) \ v ≈ Float64.(Matrix(C)) \ v
 
     # Test various linear algebra operations with decomposition
-    @test det(eigen(C)) ≈ det(C)
-    @test logdet(eigen(C)) ≈ log(det(C))
-    @test Matrix(inv(eigen(C))) ≈ inv(C)
+    println("1")
+    C_dense = Float64.(Matrix(C))
+    println("2")
+    @test det(eigen(C)) ≈ det(C_dense)
+    @test logdet(eigen(C)) ≈ logdet(C_dense)
+    @test Matrix(inv(eigen(C))) ≈ inv(C_dense)
+end
+
+@testset "eigen" begin
+    rng, P, Q, R = MersenneTwister(123456), 3, 5, 7
+
+    # Generate some positive definite matrices so that logdet can be tested.
+    A_, B_, C_ = randn(rng, P, P), randn(rng, Q, Q), randn(rng, R, R)
+    A, B, C = Symmetric(A_ * A_' + I), Symmetric(B_ * B_' + I), Symmetric(C_ * C_' + I)
+    
+    D = kronecker(A, B)
+    @show size(D), size(kronecker(D, C)), size(kronecker(kronecker(A, B), kronecker(B, A)))
+    @show typeof(eigen(D))
+    eigen_tests(rng, D)
+    eigen_tests(rng, kronecker(D, C))
+    eigen_tests(rng, kronecker(C, D))
+    eigen_tests(rng, kronecker(kronecker(A, B), kronecker(B, A)))
+    # eigen_tests(rng, kronecker(kronecker(D, C), kronecker(C, D)))
 end
