### 2019-07-28 3df4f4d
:wrench: promote typing `KroneckerProduct`
