### 2019-07-28 5d6aeba
:clock9: update benchmark to use `@belapsed`
