### 2019-07-28 2712cb4
Add Random as test-specific dependency
Testing fails unless Random is added to Project and Manifest files.
