### 2019-07-28 60152d0
Removed printing in tests
