### 2019-07-28 42b2b59
Define struct for Kronecker sum
Store terms as lazy products, but construct it from the original factor 
matrices
