### 2019-07-28 fb7b770
Define constructors
Uses similar methods as base.jl since kronecker and kroneckersum use the 
same factor-matrices.
