### 2019-07-28 1da1bf4
Define powers and binary operator for kronecker sums
The same exponentiation rules apply as in src/base.jl.
Could perhaps later be expanded to non-negative powers, or all integers 
(identity product and inverse products of invidivual terms)?
