+++ src/Kronecker.jl
@@ -2,8 +2,9 @@ module Kronecker
 
 # TODO types!
 export GeneralizedKroneckerProduct, AbstractKroneckerProduct, KroneckerProduct, SquareKroneckerProduct, EigenKroneckerProduct, ShiftedKroneckerProduct
+export AbstractKroneckerSum, KroneckerSum
 export issquare, getmatrices, size, getindices, order, issymmetric
-export ⊗, kronecker, Matrix
+export ⊗, kronecker, Matrix, ⊕, kroneckersum
 export tr, det, collect, inv, *, mult!, eigen, \, /, adjoint, transpose, conj, solve
 export getindex
 
@@ -16,5 +17,6 @@ include("base.jl")
 include("indexedkroncker.jl")
 include("shiftedkronecker.jl")
 include("eigen.jl")
+include("kroneckersum.jl")
 
 end # module
