### 2019-07-28 55c0e2c
Added tests, defined abstract types and getmatrices. Included file.
Added abstract supertype AbstractKroneckerSum.

Fixed order(M::AbstractKroneckerSum) where it now only checks the order 
of a single term in the sum.

Defined getmatrices for sums. Currently only works for sums of order 2.
