+++ test/runtests.jl
@@ -1,7 +1,7 @@
 using Kronecker
 using Test
 using LinearAlgebra
-
 include("testbase.jl")
 include("testindexed.jl")
 #include("testshifted.jl")
+include("kroneckersum_tests.jl")
