+++ src/Kronecker.jl
@@ -2,8 +2,9 @@ module Kronecker
 
 export GeneralizedKroneckerProduct, AbstractKroneckerProduct, AbstractSquareKronecker
 export SquareKroneckerProduct, EigenKroneckerProduct, ShiftedKroneckerProduct
+export AbstractKroneckerSum, KroneckerSum
 export issquare, getmatrices, size, getindices, order, issymmetric, isposdef
-export ⊗, kronecker, Matrix
+export ⊗, kronecker, Matrix, ⊕, kroneckersum
 export tr, det, logdet, collect, inv, +, *, mult!, eigen, \, /, adjoint, transpose, conj, solve
 export getindex
 export cholesky, CholeskyKronecker
@@ -18,5 +19,6 @@ include("base.jl")
 include("indexedkroncker.jl")
 include("eigen.jl")
 include("factorization.jl")
+include("kroneckersum.jl")
 
 end # module
