### 2019-07-28 66e5a4c
Defined size, getindex and eltype. Added tests.
