+++ test/kroneckersum_tests.jl
@@ -32,4 +32,16 @@
     @test order(KS) == 2
     @test order(KS3) == 3
 
+    @test getmatrices(KS) == (A,B)
+    @test_broken getmatrices(KS3) == (A,B,C)
+
+    @test getindex(KS,2,3) == kronsum[2,3]
+    @test getindex(KS3,2,3) == kronsum3[2,3]
+
+    D = rand(ComplexF64,6,6)
+    @test size(A ⊕ D) == size(A) .* size(D,1)
+    @test size(B ⊕ C ⊕ D) == size(B) .* size(C) .* size(D)
+    @test eltype(A ⊕ B) == Float64
+    @test eltype(C ⊕ D) == ComplexF64
+
 end
