### 2019-07-28 da7d8ba
Add basic LinearAlgebra functions on Kronecker sums
