### 2019-07-29 722e4fe
Added multiplication methods
