### 2019-07-29 d6a8fb5
Added multiplication methods
