### 2019-07-29 98ba671
Remove ShiftedKronecker and tidy tests
