### 2019-07-29 94da30d
Uncomment tests
