### 2019-07-29 2990e27
Removed mixed product of sums.
