### 2019-07-29 7a844ae
:wrench: restructuring KronSum
