+++ test/runtests.jl
@@ -2,6 +2,9 @@ using Kronecker
 using Test
 using LinearAlgebra
 
-include("testbase.jl")
-include("testindexed.jl")
+#include("testbase.jl")
+#include("testindexed.jl")
+#("testeigen.jl")
+include("testkroneckersum.jl")
+
 #include("testshifted.jl")
