### 2019-07-29 ee228c3
:wrench: restructure test suite
