+++ test/runtests.jl
@@ -2,9 +2,10 @@ using Kronecker
 using Test
 using LinearAlgebra
 
-#include("testbase.jl")
-#include("testindexed.jl")
-#("testeigen.jl")
+include("testbase.jl")
+include("testindexed.jl")
+include("testeigen.jl")
 include("testkroneckersum.jl")
+include("testfactorization.jl")
 
 #include("testshifted.jl")
