### 2019-07-29 fd7aa95
:bulb: adding base documentation file
