+++ docs/make.jl
@@ -1,0 +1,3 @@
+using Documenter, Kronecker
+
+makedocs(sitename="Kronecker.jl")
