+++ docs/src/index.md
@@ -1,0 +1,5 @@
+# Example.jl Documentation
+
+```@docs
+kronecker
+```
