### 2019-07-29 7711bef
:put_litter_in_its_place: remove tests shiftedkron
