+++ test/runtests.jl
@@ -7,5 +7,3 @@ include("testindexed.jl")
 include("testeigen.jl")
 include("testkroneckersum.jl")
 include("testfactorization.jl")
-
-#include("testshifted.jl")
