### 2019-07-30 615931b
:put_litter_in_its_place: removed SquareKronecker
