### 2019-07-30 090c9da
:zap: performance collect Kronecker sum
