### 2019-07-30 3ef7715
:zap: performance collect Kronecker sum
