+++ src/kroneckersum.jl
@@ -109,7 +109,7 @@ native matrix. Equivalent with `Matrix(K::AbstractKroneckerProduct)`.
 function Base.collect(K::AbstractKroneckerSum)
     A, B = getmatrices(K)
     IA, IB = oneunit(A), oneunit(B)
-    return kron(A, IB) + kron(IA, B)
+    return kron(collect(A), IB) + kron(IA, collect(B))
 end
 
 function Base.adjoint(K::AbstractKroneckerSum)
