### 2019-07-30 fdfedeb
:zap: performance higherorder Kron sum
