### 2019-07-30 1e78e35
:zap: `collect(K::AbstractKroneckerSum)` return sparse
