+++ src/kroneckersum.jl
@@ -100,13 +100,14 @@ function LinearAlgebra.tr(K::AbstractKroneckerSum)
 end
 
 """
-    collect(K::AbstractKroneckerProduct)
+    collect(K::AbstractKroneckerSum)
 
-Collects a lazy instance of the `AbstractKroneckerProduct` type into a full,
-native matrix. Equivalent with `Matrix(K::AbstractKroneckerProduct)`.
+Collects a lazy instance of the `AbstractKroneckerSum` type into a full,
+native matrix. Returns the result as a sparse matrix.
 """
 function Base.collect(K::AbstractKroneckerSum)
     A, B = getmatrices(K)
+    A, B = sparse(A), sparse(B)
     IA, IB = oneunit(A), oneunit(B)
     return kron(A, IB) + kron(IA, B)
 end
