### 2019-07-30 1db2fe8
:white_check_mark: add test for KroneckerSum types
