### 2019-07-30 6fef560
Added willtebbutt's suggestions to MichielStock/Kronecker.jl#28
