### 2019-07-30 e3d8f5c
Fixed squarecheck definition
