### 2019-07-30 caedcd0
Use `collect` in nested krons and add syntax test
