### 2019-07-30 2f716b4
Add issymmetric tests
