### 2019-07-30 237c79c
Update Kronecker.jl
Removing exporting Kronecker types
