+++ src/Kronecker.jl
@@ -1,8 +1,5 @@
 module Kronecker
 
-export GeneralizedKroneckerProduct, AbstractKroneckerProduct, AbstractSquareKronecker
-export AbstractKroneckerSum, KroneckerSum
-export EigenKroneckerProduct, ShiftedKroneckerProduct
 export issquare, getmatrices, size, getindices, order, issymmetric, isposdef
 export ⊗, kronecker, Matrix, ⊕, kroneckersum
 export tr, det, logdet, collect, inv, +, *, mult!, eigen, /, adjoint, transpose, conj, solve, exp, kron
