### 2019-07-30 5c107e5
Update Kronecker.jl
