### 2019-07-30 9e3d25c
inline functions
