### 2019-07-30 c6e36ba
inlining function
