+++ test/testbase.jl
@@ -73,7 +73,7 @@
         @test kron(A ⊗ B, C) ≈ kron(A, B, C)
         @test kron(A, B ⊗ C) ≈ kron(A, B, C)
         @test kron(A ⊗ B, C ⊗ D) ≈ kron(A, B, C, D)
-        @test collect((A⊗B) ⊗ (C⊗D)) ≈ kron(A, B, C, D))
+        @test collect((A⊗B) ⊗ (C⊗D)) ≈ kron(A, B, C, D)
     end
 
     @testset "Mixed product" begin
