### 2019-07-30 997235b
rename mul! + add test
