### 2019-08-02 b702715
:wrench: inline `kron`
