### 2019-08-02 52c378b
:white_check_mark: add error tests
