+++ test/testbase.jl
@@ -56,6 +56,10 @@
 
     @testset "Vec trick" begin
         @test K * v ≈ X * v
+        V = rand(4, 3)
+        @test K * vec(V) ≈ collect(K) * vec(V)
+        @test_throws DimensionMismatch K * V
+        @test_throws DimensionMismatch K * reshape(V, 2, 6)
         K3 = A ⊗ B ⊗ C
         v3 = randn(size(K3, 2))
         @test K3 * v3 ≈ collect(K3) * v3
