+++ test/runtests.jl
@@ -3,6 +3,7 @@ using SparseArrays: SparseMatrixCSC
 
 @testset "Kronecker" begin
     include("testbase.jl")
+    include("testvectrick.jl")
     include("testindexed.jl")
     include("testeigen.jl")
     include("testkroneckersum.jl")
