### 2019-08-04 1aeaf10
:white_check_mark: Test reshaped mul!
