+++ test/testvectrick.jl
@@ -18,3 +18,19 @@
     u = similar(v)
     @test mul!(u, K, v) ≈ X * v
 end
+
+@testset "Reshaped vec trick" begin
+    A = rand(3, 3)
+    B = ones(4, 4)
+    C = randn(5, 6)
+    K = A ⊗ B
+    X = collect(K)
+    v = rand(12)
+
+    @test K * v ≈ X * v
+
+    V = sprand(4, 3, 1.0)
+    @test K * vec(V) ≈ X * vec(V)
+    @test_throws DimensionMismatch K * V
+    @test_throws DimensionMismatch K * reshape(V, 2, 6)
+end
