### 2019-08-04 43ef438
:white_check_mark: Test mul-reshape for ReshapedArray
