+++ src/Kronecker.jl
@@ -7,6 +7,8 @@ export issquare, order, issymmetric, isposdef, getmatrices
 export ⊗, kronecker, ⊕, kroneckersum
 export CholeskyKronecker
 
+export isprob, naivesample
+
 using LinearAlgebra
 import LinearAlgebra: mul!
 import Base: collect
@@ -18,5 +20,6 @@ include("indexedkroncker.jl")
 include("eigen.jl")
 include("factorization.jl")
 include("kroneckersum.jl")
+include("kroneckergraphs.jl")
 
 end # module
