### 2019-08-08 545159b
:tada: scalar multiplication
