+++ src/vectrick.jl
@@ -75,3 +75,10 @@ function Base.:*(K::AbstractKroneckerProduct, v::AbstractVector)
     return mul!(Vector{promote_type(eltype(v), eltype(K))}(undef,
                                                         first(size(K))), K, v)
 end
+
+function Base.sum(K::AbstractKroneckerProduct)
+    A, B = getmatrices(K)
+    s = zero(eltype(K))
+    sumB = sum(B)
+    return sum(sum(B) * A)
+end
