### 2019-08-08 bf7f833
:tada: row-and col sums
