### 2019-08-09 dd854c7
:zap: fast sampling of Kronecker networks
