### 2019-08-09 9fdbf50
:heavy_plus_sign: add dependency on StatsBase
