### 2019-08-09 2c65a48
:tada: add `lmul!` and `rmul!`
