+++ src/base.jl
@@ -1,4 +1,4 @@
-abstract type GeneralizedKroneckerProduct <: AbstractMatrix{Real} end
+abstract type GeneralizedKroneckerProduct <: AbstractMatrix{Number} end
 
 abstract type AbstractKroneckerProduct <: GeneralizedKroneckerProduct end
 
@@ -234,21 +234,22 @@ function Base.:*(K1::AbstractKroneckerProduct, K2::AbstractKroneckerProduct)
     return (A * C) ⊗ (B * D)
 end
 
-```
+
+"""
     lmul!(a::Number, K::AbstractKroneckerProduct)
 
 Scale an `AbstractKroneckerProduct` `K` inplace by a factor `a` by rescaling the
 **left** matrix.
-```
+"""
 LinearAlgebra.lmul!(a::Number, K::AbstractKroneckerProduct) = lmul!(a, K.A)
 
-```
+"""
     rmul!(K::AbstractKroneckerProduct, a::Number)
 
 Scale an `AbstractKroneckerProduct` `K` inplace by a factor `a` by rescaling the
 **right** matrix.
-```
+"""
 LinearAlgebra.rmul!(K::AbstractKroneckerProduct, a::Number) = rmul!(K.B, a)
 
 Base.:*(a::Number, K::AbstractKroneckerProduct) = kronecker(a * K.A, K.B)
-Base.:*(K::AbstractKroneckerProduct, a::Number) = kronecker(K.A, K.B * a)
+Base.:*(K::AbstractKroneckerProduct, a::Number) = kronecker(a * K.A, K.B)
