### 2019-08-12 13504b4
:white_check_mark: add unit tests
