### 2019-08-12 5d648df
:white_check_mark: test `lmul!` and `rmul!`
