### 2019-08-12 ef75d41
:tadat: KroneckerPower type
