### 2019-08-14 5486561
:wrench: indexing KroneckerPower
