+++ src/Kronecker.jl
@@ -2,6 +2,7 @@ module Kronecker
 
 export GeneralizedKroneckerProduct, AbstractKroneckerProduct
 export AbstractKroneckerSum, KroneckerSum
+export KroneckerPower
 export EigenKroneckerProduct
 export issquare, order, issymmetric, isposdef, getmatrices
 export ⊗, kronecker, ⊕, kroneckersum
@@ -15,6 +16,7 @@ import Base: collect
 using SparseArrays: sparse
 
 include("base.jl")
+include("kroneckerpowers.jl")
 include("vectrick.jl")
 include("indexedkroncker.jl")
 include("eigen.jl")
