+++ src/base.jl
@@ -14,6 +14,85 @@ struct KroneckerProduct{T,TA<:AbstractMatrix, TB<:AbstractMatrix} <: AbstractKro
 end
 
 """
+    kronecker(A::AbstractMatrix, B::AbstractMatrix)
+
+Construct a Kronecker product object between two arrays. Does not evaluate the
+Kronecker product explictly.
+"""
+kronecker(A::AbstractMatrix, B::AbstractMatrix) = KroneckerProduct(A, B)
+
+"""
+    kronecker(A::AbstractMatrix, B::AbstractMatrix)
+
+Higher-order Kronecker lazy kronecker product, e.g.
+```
+kronecker(A, B, C, D)
+```
+"""
+kronecker(A::AbstractMatrix, B::AbstractMatrix...) = kronecker(A,
+                                                            kronecker(B...))
+"""
+    kronecker(A::AbstractMatrix, pow::Int)
+
+Kronecker power, computes `A ⊗ A ⊗ ... ⊗ A`.
+"""
+function kronecker(A::AbstractMatrix, pow::Int)
+    @assert pow > 0 "Works only with positive powers!"
+    if pow == 1
+        return A
+    else
+        return A ⊗ kronecker(A, pow-1)
+    end
+end
+
+"""
+    ⊗(A::AbstractMatrix, B::AbstractMatrix)
+
+Binary operator for `kronecker`, computes as Lazy Kronecker product. See
+`kronecker` for documentation.
+"""
+⊗(A::AbstractMatrix, B::AbstractMatrix) = kronecker(A, B)
+⊗(A::AbstractMatrix...) = kronecker(A...)
+⊗(A::AbstractMatrix, pow::Int) = kronecker(A, pow)
+
+function Base.getindex(K::AbstractKroneckerProduct, i1::Int, i2::Int)
+    A, B = getmatrices(K)
+    m, n = size(A)
+    k, l = size(B)
+    return A[cld(i1, k), cld(i2, l)] * B[(i1 - 1) % k + 1, (i2 - 1) % l + 1]
+end
+
+"""
+    getmatrices(K::AbstractKroneckerProduct)
+
+Obtain the two matrices of an `AbstractKroneckerProduct` object.
+"""
+getmatrices(K::AbstractKroneckerProduct) = (K.A, K.B)
+
+"""
+    getmatrices(A::AbstractArray)
+
+Returns a matrix itself. Needed for recursion.
+"""
+getmatrices(A::AbstractArray) = (A,)
+
+function Base.eltype(K::AbstractKroneckerProduct)
+    A, B = getmatrices(K)
+    return promote_type(eltype(A), eltype(B))
+end
+
+function Base.size(K::AbstractKroneckerProduct)
+    A, B = getmatrices(K)
+    (m, n) = size(A)
+    (k, l) = size(B)
+    return m * k, n * l
+end
+
+Base.size(K::GeneralizedKroneckerProduct, dim::Int) = size(K)[dim]
+
+# CHECKS
+
+"""
     issquare(A::AbstractMatrix)
 
 Checks if an array is a square matrix.
@@ -45,6 +124,19 @@ function LinearAlgebra.issymmetric(K::AbstractKroneckerProduct)
 end
 
 """
+    isposdef(K::AbstractKroneckerProduct)
+
+Test whether a Kronecker product is positive definite (and Hermitian) by trying
+to perform a Cholesky factorization of K.
+"""
+function LinearAlgebra.isposdef(K::AbstractKroneckerProduct)
+    squarecheck(K)
+    return isposdef(K.A) && isposdef(K.B)
+end
+
+# SCALAR PROPERTIES
+
+"""
     order(M::AbstractMatrix)
 
[…]
