### 2019-08-14 6b955a1
:tada: kronecker powers functionality!
