### 2019-08-14 9ae2b1d
:white_check_mark: tests form Kronecker powers
