### 2019-08-16 a626a84
:ambulance: fix error sampling Kronecker powers
