+++ test/testkroneckergraphs.jl
@@ -1,4 +1,4 @@
-@test "Kronecker graphs" begin
+@testset "Kronecker graphs" begin
     P1 = [0.1 0.0 0.0;
           0.3 0.4 0.5;
           1.0 0.7 0.2]
@@ -24,7 +24,7 @@
           @test G isa AbstractSparseMatrix
           @test size(G) == size(P)
           @test G[9, 1]  # probability of 1
-          @test !G[2, 5]
+          @test !G[2, 5]  # probability of 0
     end
 
     @testset "helpers" begin
@@ -33,22 +33,22 @@
          @test isprob(P)
          @test isprob(P4order)
 
-         I = sampleindices(P1, 100)
-         @test all(maximum(I) .<= size(P1))
-         for ind in I
-               @test P1[I] > 0
+         indices = sampleindices(P1, 100)
+         @test all(maximum(indices) .<= size(P1))
+         for ind in indices
+               @test P1[ind...] > 0
          end
 
-         I = sampleindices(P, 100)
-         @test all(maximum(I) .<= size(P))
-         for ind in I
-               @test P[I] > 0
+         indices = sampleindices(P, 100)
+         @test all(maximum(indices) .<= size(P))
+         for ind in indices
+               @test P[ind...] > 0
          end
 
-         I = sampleindices(P4order, 1000)
-         @test all(maximum(I) .<= size(P4order))
-         for ind in I
-               @test P4order[I] > 0
+         indices = sampleindices(P4order, 1000)
+         @test all(maximum(indices) .<= size(P4order))
+         for ind in indices
+               @test P4order[ind...] > 0
          end
    end
 
