+++ test/runtests.jl
@@ -1,5 +1,5 @@
 using Kronecker, Test, LinearAlgebra
-using SparseArrays: SparseMatrixCSC, sprand
+using SparseArrays: SparseMatrixCSC, sprand, AbstractSparseMatrix
 
 @testset "Kronecker" begin
     include("testbase.jl")
@@ -9,4 +9,5 @@ using SparseArrays: SparseMatrixCSC, sprand
     include("testeigen.jl")
     include("testkroneckersum.jl")
     include("testfactorization.jl")
+    include("testkroneckergraphs.jl")
 end
