### 2019-08-16 d1d5a61
:ambulance: eigen for Kronecker powers
