+++ test/testvectrick.jl
@@ -45,4 +45,6 @@ end
 
     @test sum(K, dims=1) ≈ sum(X, dims=1)
     @test sum(K3, dims=2) ≈ sum(collect(K3), dims=2)
+
+    @test sum(kronecker(A, 3)) ≈ sum(kron(A, A, A))
 end
