### 2019-08-20 #15 @matbesancon
ping @mcognetta to continue the review
