### 2019-08-23 6e1d5f7
:pencil2: document `conj`
