### 2019-08-30 #15 @mcognetta
In light of the revisions, I am changing my recommendation to **accept as-is.**
