### 2019-09-01 93d131f
Resolve code duplication issue
