### 2019-09-01 5cd55ff
Remove comments and fix style
