+++ src/vectrick.jl
@@ -50,6 +50,36 @@ function Base.:*(K::AbstractKroneckerProduct, v::AbstractVector)
     return mul!(Vector{promote_type(eltype(v), eltype(K))}(undef, first(size(K))), K, v)
 end
 
+reshape_cols(x, sizes...) = reshape(x, sizes...)
+reshape_rows(x, sizes...) = transpose(reshape(collect(transpose(x)), reverse(sizes)...))
+
+kron_id_a(a, x) = reshape_cols(a * reshape_cols(x, size(a)[2], :), :, size(x)[2])
+kron_a_id(a, x) = reshape_rows(a * reshape_rows(x, size(a)[2], :), :, size(x)[2])
+function kron_a_b(A, B, x)
+    a, b = size(A)
+    c, d = size(B)
+    if (d + a) * b < (b + c) * d
+        return kron_a_id(A, kron_id_a(B, x))
+    else
+        return kron_id_a(B, kron_a_id(A, x))
+    end
+end
+
+function Base.:*(A::AbstractKroneckerProduct, B::StridedMatrix)
+    size(A, 2) != size(B, 1) && throw(DimensionMismatch("size(A, 2) != size(B, 1)"))
+    return kron_a_b(getmatrices(A)..., B)
+end
+
+function Base.:*(A::KroneckerProduct{<:Any, <:Eye, <:AbstractMatrix}, B::StridedMatrix)
+    size(A, 2) != size(B, 1) && throw(DimensionMismatch("size(A, 2) != size(B, 1)"))
+    return kron_id_a(A.B, B)
+end
+
+function Base.:*(A::KroneckerProduct{<:Any, <:AbstractMatrix, <:Eye}, B::StridedMatrix)
+    size(A, 2) != size(B, 1) && throw(DimensionMismatch("size(A, 2) != size(B, 1)"))
+    return kron_a_id(A.A, B)
+end
+
 function Base.sum(K::AbstractKroneckerProduct; dims::Union{Nothing,Int}=nothing)
     A, B = getmatrices(K)
     if dims == nothing
