### 2019-09-02 97f658c
Improve perf + add tests
