+++ docs/src/man/linalgebra.md
@@ -1,0 +1,11 @@
+# Linear Algebra
+
+```@docs
+det(K::AbstractKroneckerProduct)
+logdet(K::AbstractKroneckerProduct)
+tr(K::AbstractKroneckerProduct)
+inv(K::AbstractKroneckerProduct)
+adjoint(K::AbstractKroneckerProduct)
+transpose(K::AbstractKroneckerProduct)
+conj(K::AbstractKroneckerProduct)
+```
