### 2019-09-04 344a6c2
:pencil2: update Kronecker paper
