+++ paper/paper.tex
@@ -12,28 +12,28 @@
 
 \begin{abstract}
 Pairwise learning is a machine learning paradigm where the goal is to predict properties of pairs of objects.
-Applications include recommender systems, molecular network inference and ecological interaction prediction.
-Kronecker-based learning systems provide a simple, yet elegant method to learn from such pairs.
-Using tricks from linear algebra, these models can be trained, tuned and validated on large datasets.
-Our Julia package \texttt{Kronecker.jl} aggregates these tricks using a lazily-evaluated Kronecker product $\otimes$, such that it is easy to experiment with learning algorithms using the Kronecker product.
+Applications include recommender systems, molecular network inference, and ecological interaction prediction.
+Kronecker-based learning systems provide a simple yet elegant method to learn from such pairs.
+Using tricks from linear algebra, these models can be trained, tuned, and validated on large datasets.
+Our Julia package \texttt{Kronecker.jl} aggregates these shortcuts and efficient algorithms using a lazily-evaluated Kronecker product `$\otimes$', such that it is easy to experiment with learning algorithms using the Kronecker product.
 
 \end{abstract}
 
 \section{Background}
 
-In linear algebra, the Kronecker product, denoted by $\otimes$, between an $(n\times m)$ matrix $A=[A_{ij}]$ and an $(p\times q)$ matrix $B=[B_{kl}]$ is computed as
+The Kronecker product, denoted by $\otimes$, between an $(n\times m)$ matrix $A=[A_{ij}]$ and an $(p\times q)$ matrix $B=[B_{kl}]$ is computed as
 \begin{equation}\label{eq:kron}
   {A} \otimes  {B} ={\begin{bmatrix}A_{11} {B} &\cdots &A_{1m} {B} \\\vdots &\ddots &\vdots \\A_{n1} {B} &\cdots &A_{nm} {B} \end{bmatrix}}\,.
 \end{equation}
 Simply put, the Kronecker product creates a new $(np\times mq)$ matrix containing all element-wise products between the respective elements of the two matrices.
 
-Though conceptually simple, the Kronecker product gives rise to some some elegant mathematics~\cite{Sch2013,VanLoan2000}.
-The Kronecker product has numerous applications in applied mathematics, for example in defining the matrix normal distribution, modelling complex networks~\cite{Leskovec2008} and pairwise learning~\cite{Stock2017tskrr}.
-The reason that it can be used in practice is that (\ref{eq:kron}) often does not have to be computed explicitly, but it can be circumvented using various computational shortcuts.
+Though conceptually simple, the Kronecker product gives rise to some elegant mathematics which allows performing many important computations, such as the eigenvalue decomposition, determinant or trace, in an efficient way~\cite{Sch2013,VanLoan2000}.
+The Kronecker product has numerous applications in applied mathematics, for example in defining the matrix normal distribution, modeling complex networks~\cite{Leskovec2008} and pairwise learning~\cite{Stock2017tskrr}.
+The reason that one can use the Kronecker product in large numerical problems is that (\ref{eq:kron}) often does not have to be computed explicitly, but it can be circumvented using various computational shortcuts.
 
 \section{Basic use}
 
-Our package aims to be a toolkit to easily build Kronecker-based applications, where the focus is on the mathematics and computational efficiency is taken care of under the hood. Essentially, it provides a lazily-evaluated Kronecker product of a \texttt{Kronecker} type.
[…]
