+++ paper/ref.bib
@@ -1,3 +1,33 @@
+@article{Besard2019,
+abstract = {GPUs and other accelerators are popular devices for accelerating compute-intensive, parallelizable applications. However, programming these devices is a difficult task. Writing efficient device code is challenging, and is typically done in a low-level programming language. High-level languages are rarely supported, or do not integrate with the rest of the high-level language ecosystem. To overcome this, we propose compiler infrastructure to efficiently add support for new hardware or environments to an existing programming language. We evaluate our approach by adding support for NVIDIA GPUs to the Julia programming language. By integrating with the existing compiler, we significantly lower the cost to implement and maintain the new compiler, and facilitate reuse of existing application code. Moreover, use of the high-level Julia programming language enables new and dynamic approaches for GPU programming. This greatly improves programmer productivity, while maintaining application performance similar to that of the official NVIDIA CUDA toolkit.},
+archivePrefix = {arXiv},
+arxivId = {1712.03112v1},
+author = {Besard, Tim and Foket, Christophe and {De Sutter}, Bjorn},
+doi = {10.1109/TPDS.2018.2872064},
+eprint = {1712.03112v1},
+file = {:Users/michielstock/Library/Application Support/Mendeley Desktop/Downloaded/Besard, Foket, Sutter - Unknown - IEEE TRANSACTIONS ON PARALLEL AND DISTRIBUTED SYSTEMS 1 Effective Extensible Programming Unleashing Ju.pdf:pdf},
+issn = {15582183},
+journal = {IEEE Transactions on Parallel and Distributed Systems},
+keywords = {Graphics processors,code generation,retargetable compilers,very high-level languages},
+number = {4},
+pages = {827--841},
+title = {{Effective extensible programming: unleashing Julia on GPUs}},
+url = {https://arxiv.org/pdf/1712.03112.pdf},
+volume = {30},
+year = {2019}
+}
+@article{Innes2019,
+abstract = {Scientific computing is increasingly incorporating the advancements in machine learning and the ability to work with large amounts of data. At the same time, machine learning models are becoming increasingly sophisticated and exhibit many features often seen in scientific computing, stressing the capabilities of machine learning frameworks. Just as the disciplines of scientific computing and machine learning have shared common underlying infrastructure in the form of numerical linear algebra, we now have the opportunity to further share new computational infrastructure, and thus ideas, in the form of Differentiable Programming. We describe Zygote, a Differentiable Programming system that is able to take gradients of general program structures. We implement this system in the Julia programming language. Our system supports almost all language constructs (control flow, recursion, mutation, etc.) and compiles high-performance code without requiring any user intervention or refactoring to stage computations. This enables an expressive programming model for deep learning, but more importantly, it enables us to incorporate a large ecosystem of libraries in our models in a straightforward way. We discuss our approach to automatic differentiation, including its support for advanced techniques such as mixed-mode, complex and checkpointed differentiation, and present several examples of differentiating programs.},
+archivePrefix = {arXiv},
+arxivId = {1907.07587},
+author = {Innes, Mike and Edelman, Alan and Fischer, Keno and Rackauckas, Chris and Saba, Elliot and Shah, Viral B and Tebbutt, Will},
+eprint = {1907.07587},
+file = {:Users/michielstock/Library/Application Support/Mendeley Desktop/Downloaded/Innes et al. - Unknown - ∂P A Differentiable Programming System to Bridge Machine Learning and Scientific Computing.pdf:pdf},
+journal = {Innes, M Edelman, A Fischer, K Rackauckas, C Saba, E Shah, Viral Tebbutt, W},
[…]
