+++ paper/paper.yml
@@ -19,5 +19,5 @@ affiliations:
     index: 1
   - name:  Department of Future Technologies
     index: 2
-date: 23 June 2019
+date: 4 September 2019
 bibliography: ref.bib
