+++ paper/ref.bib
@@ -23,7 +23,6 @@ arxivId = {1907.07587},
 author = {Innes, Mike and Edelman, Alan and Fischer, Keno and Rackauckas, Chris and Saba, Elliot and Shah, Viral B and Tebbutt, Will},
 eprint = {1907.07587},
 file = {:Users/michielstock/Library/Application Support/Mendeley Desktop/Downloaded/Innes et al. - Unknown - ∂P A Differentiable Programming System to Bridge Machine Learning and Scientific Computing.pdf:pdf},
-journal = {Innes, M Edelman, A Fischer, K Rackauckas, C Saba, E Shah, Viral Tebbutt, W},
 title = {{A differentiable programming system to bridge machine learning and scientific computing}},
 url = {https://arxiv.org/pdf/1907.07587.pdf http://arxiv.org/abs/1907.07587},
 year = {2019}
