### 2019-09-04 #15 @MichielStock
These comments are now done as of d33dc39371e26083b43f955df7855879884e449c

> On 3 Sep 2019, at 17:05, Mathieu Besançon <notifications@github.com> wrote:
>
> @MichielStock <https://github.com/MichielStock> a couple additional comments:
> Section 1: "In linear algebra" -> is it necessary, given that the context is operations over matrices?
>
> "gives rises to some elegant mathematics", elegant is a bit odd to see (hard to define). To some properties of interest?
>
> "The elementary functions of LinearAlgebra" -> "The elementary functions of the LinearAlgebra module from the Julia standard library".
>
> I would indicate more clearly that "vec trick" is a standard term in the literate, by using quotes and hints like "so-called vec trick" with the reference to [5]. I would not use "tricks from linear algebra" in the abstract as it sounds a bit informal.
>
> Prospects
> Could you cite the packages you are referencing by adding a bibtex entry? (Zygote, CuArrays), also: CuArrays -> CuArrays.jl
>
