### 2019-09-05 f8c0b2e
:put_litter_in_its_place: removing needless computation in sum
