+++ paper/paper.tex
@@ -23,7 +23,7 @@ Our Julia package \texttt{Kronecker.jl} aggregates these shortcuts and efficient
 
 The Kronecker product, denoted by $\otimes$, between an $(n\times m)$ matrix $A=[A_{ij}]$ and an $(p\times q)$ matrix $B=[B_{kl}]$ is computed as
 \begin{equation}\label{eq:kron}
-  {A} \otimes  {B} ={\begin{bmatrix}A_{11} {B} &\cdots &A_{1m} {B} \\\vdots &\ddots &\vdots \\A_{n1} {B} &\cdots &A_{nm} {B} \end{bmatrix}}\,.
+  {A} \otimes  {B} ={\begin{bmatrix}A_{1,1} {B} &\cdots &A_{1,m} {B} \\\vdots &\ddots &\vdots \\A_{n,1} {B} &\cdots &A_{n,m} {B} \end{bmatrix}}\,.
 \end{equation}
 Simply put, the Kronecker product creates a new $(np\times mq)$ matrix containing all element-wise products between the respective elements of the two matrices.
 
@@ -43,8 +43,11 @@ K = kronecker(A, B)  # lazy Kronecker product
 \end{lstlisting}
 %\end{verbatim}
 
-Alternatively, one can make use of Julia's Unicode support: \texttt{K = A $\otimes$ B}.
-The elementary functions of \texttt{LinearAlgebra} are overloaded to work with the \texttt{Kronecker} type and are all fast.
+Alternatively, one can make use of Julia's Unicode support:
+$$
+\text{\texttt{K = A $\otimes$ B}}\,.
+$$
+The elementary functions of \texttt{LinearAlgebra} are overloaded to work with the respective subtypes of \texttt{GeneralizedKroneckerProduct} and provide the most efficient implementation.
 
 \begin{lstlisting}[language = Julia]
 tr(K)  # computed as tr(A) * tr(B)
@@ -56,7 +59,10 @@ K * v  # computed using the vec trick
 \end{lstlisting}
 
 For example, the last line is evaluated using the so-called "vec trick"~\cite{VanLoan2000} with a time complexity of $\mathcal{O}(nm+pq)$ instead of $\mathcal{O}(nmpq)$ naively.
-Similarly, efficiently solving large shifted Kronecker systems can be done directly as \texttt{(A $\otimes$ B +$\lambda$I) $\backslash$ v}.
+Similarly, efficiently solving large shifted Kronecker systems can be done directly as \texttt{eigen(A $\otimes$ B +$\lambda$I) $\backslash$ v}, exploiting the fast eigenvalue decomposition for Kronecker products.
+
+Our package fully supports higher-order Kronecker products, e.g. \texttt{A $\otimes$ B $\otimes$ C}.
+The structure \texttt{KroneckerPower} (for example constructed as \texttt{kronecker(A, 4)} for $A \otimes A \otimes A \otimes A$) and its methods provide efficient storage and manipulation of repeated Kronecker multiplications of the same matrix. We also provide the functionality to generate Kronecker graphs.
 
 We provide support for dealing with submatrices of a Kronecker product through the sampled vec trick~\cite{Airola2017genvectric}.
 
@@ -69,12 +75,13 @@ u = randn(100);
 Ksubset * u  # computed using sampled vec trick
 \end{lstlisting}
 
-\texttt{Kronecker.jl} is a package in development. The developers are continuously adding new features.
-
 \section{Prospects}
 
[…]
