### 2019-09-05 #15 @MichielStock
Updated. Also added a section on higher-order Kronecker products, which are now supported.

> On 5 Sep 2019, at 14:52, Mathieu Besançon <notifications@github.com> wrote:
>
> Alternatively, one can make use of Julia's Unicode support: K = A \cross B
>
> Can you make K = A cross B an equation as to have it all on the same line?
>
> overloaded to work with the Kronecker type.
>
> As I understand, there are several Kronecker types, should it be pluralized?
>
> The last sentence of section 2 feels like a conclusion and prospect, should it go at the beginning of section 3?
>
> At the end of section 3, could you add some more details on the extensions for high-order Kronecker products? (not a big deal if the text is a bit above 1 page)
>
