+++ test/cholesky.jl
@@ -19,8 +19,6 @@ to_psd(A) = A * A' + I
     @test det(chol_A_kron_B) ≈ det(chol_A_kron_B_dense)
     @test logdet(chol_A_kron_B) ≈ logdet(chol_A_kron_B_dense)
 
-    @show typeof(chol_A_kron_B.U)
-    @show typeof(chol_A_kron_B.U')
 
     # Test backsolve vs dense vector from the left.
     x = randn(rng, M * N)
