### 2019-09-10 b92a159
:ambulance: fix solving Cholesky using easy inverting of matrices
