### 2019-09-10 87081bc
reduce allocations
