### 2019-09-12 1cf62a0
:wrench: propagate check in `cholesky`
