+++ docs/make.jl
@@ -1,3 +1,5 @@
 using Documenter, Kronecker, LinearAlgebra
 
-makedocs(sitename="Kronecker.jl", modules = [Kronecker])
+makedocs(sitename="Kronecker.jl",
+        authors = "Michiel Stock",
+        modules = [Kronecker])
