+++ docs/src/index.md
@@ -1,10 +1,9 @@
-# Kronecker.jl Documentation
+# Kronecker.jl
 
 ```@contents
 Pages = [
     "man/basic.md",
     "man/types.md",
-    "man/linalgebra.md",
     "man/multiplication.md"
     "man/factorization.md",
     "man/kroneckersums.md",
@@ -19,90 +18,42 @@ Pages = [
 ## Package features
 
 - `tr`, `det`, `size`, `eltype`, `inv`, ... are efficient functions to work with Kronecker products. Either the result is a numeric value, or returns a new `KroneckerProduct` type.
-- Kronecker product - vector multiplications are performed using the vec trick.
+- Kronecker product - vector multiplications are performed using the vec trick. Two Kronecker products of conformable size can be multiplied efficiently, yielding another Kronecker product.
 - Working with incomplete systems using the [sampled vec trick](https://arxiv.org/pdf/1601.01507.pdf).
 - overloading of the the function `eigen` to compute eigenvalue decompostions of Kronecker products. Can be used to efficiently solve systems of the form `(A ⊗ B +λI) \ v`.
 - Higher-order Kronecker systems are supported: most functions work on `A ⊗ B ⊗ C` or systems of arbitrary order.
   - Efficient sampling of [Kronecker graphs](https://cs.stanford.edu/~jure/pubs/kronecker-jmlr10.pdf) is supported.
-- Kronecker powers are supported: `kronecker(A, 3)` or `A⊗3`.
+- Kronecker powers are supported: `kronecker(A, 3)` or `A ⊗ 3`.
 - A `KroneckerSum` can be constructed with `A ⊕ B` (typed using `\oplus TAB`) or `kroneckersum(A,B)`.
   - Multiplication with vectors uses  a specialized version of the vec trick
   - Higher-order sums are supported, e.g. `A ⊕ B ⊕ C` or `kroneckersum(A,4)`.
 
 ## Example use
 
-```julia
-julia> A = randn(4, 4);
+```@repl
+using Kronecker, LinearAlgebra
 
-julia> B = Array{Float64, 2}([1 2 3;
+A = [1.0 2.0;
+     3.0 5.0];
+B = Array{Float64, 2}([1 2 3;
             4 5 6;
             7 -2 9]);
 
-julia> K = A ⊗ B
-12×12 Kronecker.KroneckerProduct{Float64,Array{Float64,2},Array{Float64,2}}:
- -1.33612   -2.67224    -4.00836    1.34896    2.69793   4.04689  …  -1.71266   -2.56898   0.959048    1.9181      2.87714
- -5.34448   -6.6806     -8.01672    5.39586    6.74482   8.09378     -4.28164   -5.13797   3.83619     4.79524     5.75429
- -9.35284    2.67224   -12.0251     9.44275   -2.69793  12.1407       1.71266   -7.70695   6.71333    -1.9181      8.63143
-  0.182498   0.364996    0.547493   0.775931   1.55186   2.32779     -1.14276   -1.71414   0.0317447   0.0634894   0.0952341
-  0.729991   0.912489    1.09499    3.10373    3.87966   4.65559     -2.8569    -3.42828   0.126979    0.158724    0.190468
-  1.27748   -0.364996    1.64248    5.43152   -1.55186   6.98338  …   1.14276   -5.14241   0.222213   -0.0634894   0.285702
- -1.0181    -2.03621    -3.05431   -0.770815  -1.54163  -2.31245      0.803382   1.20507  -0.945128   -1.89026    -2.83538
[…]
