+++ docs/src/man/basic.md
@@ -1,6 +1,6 @@
 # Basic use
 
-Kronecker products between two matrices `A` and `B` by either
+Compute a lazyronecker products between two matrices `A` and `B` by either
 
 ```julia
 K = kronecker(A, B)
@@ -12,23 +12,40 @@ or using the binary operator:
 K = A ⊗ B
 ```
 
-The Kronecker product `K` behaves as a matrix, for which `size(K)`, `eltype(K)` works as expected. Elements can be accessed via `K[i,j]`, every element is computed on the fly. The function `collect` can be used to turn `K` in a regular, dense matrix.
+Note, `⊗` can be formed by typing `\otimes<tab>`.
+
+The Kronecker product `K` behaves as a matrix, for which `size(K)`, `eltype(K)` works as one would expect. Elements can be accessed via `K[i,j]`, every element is computed on the fly. The function `collect` can be used to turn `K` in a regular, dense matrix.
 
 ## Constructing Kronecker products
 
 ```@docs
 kronecker
-\otimes
-collect(K::AbstractKroneckerProduct)
+⊗
+collect(::AbstractKroneckerProduct)
 ```
 
 ## Basic properties of Kronecker products
 
 ```@docs
-getindex(K::AbstractKroneckerProduct, i1::Int, i2::Int)
-eltype(K::AbstractKroneckerProduct)
+getindex
+eltype
+size(::AbstractKroneckerProduct)
 order
 getmatrices
 issquare
 issymmetric
+```
+
+## Linear algebra
+
+Many functions of the `LinearAlgebra` module are overloaded to work with subtypes of `GeneralizedKroneckerProduct`.
+
+```@docs
+det(K::AbstractKroneckerProduct)
+logdet(K::AbstractKroneckerProduct)
+tr(K::AbstractKroneckerProduct)
+inv(K::AbstractKroneckerProduct)
+adjoint(K::AbstractKroneckerProduct)
+transpose(K::AbstractKroneckerProduct)
+conj(K::AbstractKroneckerProduct)
 ```
