### 2019-09-17 4136ec2
:pencil2: documentation Eigen
