### 2019-09-17 4c44319
:pencil2: add docs for Cholesky
