### 2019-09-17 3a4b56e
:pencil2: add docs Kronecker graphs
