### 2019-09-17 9044085
:pencil2: indexed kron
