### 2019-09-17 8cdc912
:pencil2: Kronecker sums docs
