### 2019-09-17 2e9da9e
:pencil2: typo in docstring
