+++ docs/src/man/basic.md
@@ -1,12 +1,12 @@
 # Basic use
 
-Compute a lazyronecker products between two matrices `A` and `B` by either
+Compute a lazy Kronecker products between two matrices `A` and `B` by either
 
 ```julia
 K = kronecker(A, B)
 ```
 
-or using the binary operator:
+or, by using the binary operator:
 
 ```julia
 K = A ⊗ B
@@ -14,7 +14,17 @@ K = A ⊗ B
 
 Note, `⊗` can be formed by typing `\otimes<tab>`.
 
-The Kronecker product `K` behaves as a matrix, for which `size(K)`, `eltype(K)` works as one would expect. Elements can be accessed via `K[i,j]`, every element is computed on the fly. The function `collect` can be used to turn `K` in a regular, dense matrix.
+The Kronecker product `K` behaves like a matrix, for which `size(K)`, `eltype(K)` works as one would expect. Elements can be accessed via `K[i,j]`; every element is computed on the fly. The function `collect` can be used to turn `K` in a regular, dense matrix.
+
+```@repl
+using Kronecker
+A = randn(4, 4)
+B = rand(1:10, 5, 7)
+K = A ⊗ B
+K[4, 5]
+eltype(K)  # promotion
+collect(K)
+```
 
 ## Constructing Kronecker products
 
