+++ docs/make.jl
@@ -1,3 +1,13 @@
+using Pkg
+
+tmp_packages = ["Kronecker", "Documenter"]
+
+push!(LOAD_PATH,"../src/")
+
+Pkg.activate(".")
+
+Pkg.add.(tmp_packages) # IMPORTANT
+
 using Documenter, Kronecker, LinearAlgebra
 
 makedocs(sitename="Kronecker.jl",
