### 2019-09-19 cdaed6b
:wrench: update make.jl for docs
