+++ docs/make.jl
@@ -1,5 +1,6 @@
 using Pkg
 
+#=
 tmp_packages = ["Kronecker", "Documenter"]
 
 push!(LOAD_PATH,"../src/")
@@ -7,11 +8,13 @@ push!(LOAD_PATH,"../src/")
 Pkg.activate(".")
 
 Pkg.add.(tmp_packages) # IMPORTANT
+=#
 
 using Documenter, Kronecker, LinearAlgebra
 
 makedocs(sitename="Kronecker.jl",
         authors = "Michiel Stock",
+        format = :html,
         modules = [Kronecker])
 
 deploydocs(
