+++ docs/make.jl
@@ -1,6 +1,5 @@
 using Pkg
 
-#=
 tmp_packages = ["Kronecker", "Documenter"]
 
 push!(LOAD_PATH,"../src/")
@@ -8,7 +7,6 @@ push!(LOAD_PATH,"../src/")
 Pkg.activate(".")
 
 Pkg.add.(tmp_packages) # IMPORTANT
-=#
 
 using Documenter, Kronecker, LinearAlgebra
 
