### 2019-09-19 #15 @matbesancon
Still the code snippet to turn into equation:

> Alternatively, one can make use of Julia's Unicode support: K = A \cross B

> Can you make K = A cross B an equation as to have it all on the same line?
