### 2019-09-19 af1d3ce
:pencil2: equation paper
