### 2019-09-19 #15 @MichielStock
Done in af1d3ced <https://github.com/MichielStock/Kronecker.jl/commit/af1d3cedb2e58902eead8584f3e04e9ddcab5ea1>. Would be nice to wrap this up.

> On 19 Sep 2019, at 16:40, Mathieu Besançon <notifications@github.com> wrote:
>
> Still the code snippet to turn into equation:
>
> Alternatively, one can make use of Julia's Unicode support: K = A \cross B
>
> Can you make K = A cross B an equation as to have it all on the same line?
>
