+++ docs/make.jl
@@ -1,3 +1,4 @@
+#=
 using Pkg
 
 tmp_packages = ["Kronecker", "Documenter"]
@@ -7,6 +8,7 @@ push!(LOAD_PATH,"../src/")
 Pkg.activate(".")
 
 Pkg.add.(tmp_packages) # IMPORTANT
+=#
 
 using Documenter, Kronecker, LinearAlgebra
 
