### 2019-09-27 #15 @matbesancon
ah right.
@MichielStock the next steps are: archiving the code repo with Zenodo to obtain a DOI, we'll then be able to attach the DOI to the paper
