### 2019-09-28 #15 @MichielStock
Registered to Zenodo and a badge is added to the `readme.md`: [![DOI](https://zenodo.org/badge/160198570.svg)](https://zenodo.org/badge/latestdoi/160198570)

> On 27 Sep 2019, at 19:13, Mathieu Besançon <notifications@github.com> wrote:
>
> ah right.
> @MichielStock <https://github.com/MichielStock> the next steps are: archiving the code repo with Zenodo to obtain a DOI, we'll then be able to attach the DOI to the paper
>
