### 2019-10-15 cb70bbe
:wrench: test sum kron power
