+++ test/testkroneckerpowers.jl
@@ -25,6 +25,8 @@
         @test !issymmetric(K1)
 
         @test collect(K1) ≈ K1dense
+
+        @test sum(K1) ≈ sum(K1dense)
     end
 
     @testset "Linear algebra" begin
