### 2019-10-16 4573962
:ambulance: fix sum kron graph
