### 2019-10-18 4f4560a
:tada: sum over kron sum
