### 2019-10-18 970b801
:tada: sum over dimensions for kron sum
