### 2019-11-09 ef0c697
:wrench: fixing type system KP, IKP and KronPowers
