+++ test/runtests.jl
@@ -6,8 +6,8 @@ using SparseArrays: SparseMatrixCSC, sprand, AbstractSparseMatrix
     include("testkroneckerpowers.jl")
     include("testvectrick.jl")
     include("testindexed.jl")
-    include("testeigen.jl")
-    include("testkroneckersum.jl")
+    #include("testeigen.jl")
+    #include("testkroneckersum.jl")
     include("testfactorization.jl")
     include("testkroneckergraphs.jl")
 end
