+++ src/base.jl
@@ -1,15 +1,15 @@
-abstract type GeneralizedKroneckerProduct <: AbstractMatrix{Number} end
+abstract type GeneralizedKroneckerProduct{T} <: AbstractMatrix{T} end
 
-abstract type AbstractKroneckerProduct <: GeneralizedKroneckerProduct end
+abstract type AbstractKroneckerProduct{T} <: GeneralizedKroneckerProduct{T} end
 
 Base.IndexStyle(::Type{<:GeneralizedKroneckerProduct}) = IndexCartesian()
 
 """
-    KroneckerProduct{T,TA<:AbstractMatrix, TB<:AbstractMatrix} <: AbstractKroneckerProduct
+    KroneckerProduct{T,TA<:AbstractMatrix, TB<:AbstractMatrix} <: AbstractKroneckerProduct{T}
 
 Concrete Kronecker product between two matrices `A` and `B`.
 """
-struct KroneckerProduct{T,TA<:AbstractMatrix, TB<:AbstractMatrix} <: AbstractKroneckerProduct
+struct KroneckerProduct{T<:Any,TA<:AbstractMatrix, TB<:AbstractMatrix} <: AbstractKroneckerProduct{T}
     A::TA
     B::TB
     function KroneckerProduct(A::AbstractMatrix{T}, B::AbstractMatrix{V}) where {T, V}
@@ -72,9 +72,8 @@ Returns a matrix itself. Needed for recursion.
 """
 getmatrices(A::AbstractArray) = (A,)
 
-function eltype(K::AbstractKroneckerProduct)
-    A, B = getmatrices(K)
-    return promote_type(eltype(A), eltype(B))
+function Base.eltype(K::AbstractKroneckerProduct{T}) where {T}
+    return T
 end
 
 """
