### 2019-11-09 d797bb9
:wrench: typing kron sum
