+++ src/kroneckerpowers.jl
@@ -28,7 +28,7 @@ end
 Kronecker power, computes `A ⊗ A ⊗ ... ⊗ A`. Returns a lazy `KroneckerPower`
 type.
 """
-kronecker(A::AbstractMatrix, pow::Int) = KroneckerPower(A, pow)
+kronecker(A::AbstractMatrix, pow::Integer) = KroneckerPower(A, pow)
 
 """
     ⊗(A::AbstractMatrix, pow::Int)
@@ -36,7 +36,7 @@ kronecker(A::AbstractMatrix, pow::Int) = KroneckerPower(A, pow)
 Kronecker power, computes `A ⊗ A ⊗ ... ⊗ A`. Returns a lazy `KroneckerPower`
 type.
 """
-⊗(A::AbstractMatrix, pow::Int) = kronecker(A, pow)
+⊗(A::AbstractMatrix, pow::Integer) = kronecker(A, pow)
 
 getmatrices(K::KroneckerPower{T,TA,N}) where {T,TA,N} = (K.A, KroneckerPower(K.A, K.pow-1))
 getmatrices(K::KroneckerPower{T,TA,2}) where {T,TA} = (K.A, K.A)
