### 2019-11-09 4da0a48
:white_check_mark: coverage indexed kron
