### 2019-11-09 0e62b2a
Bump version number
Type inheritance!
