+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/kroneckerpowers/#Kronecker-powers-and-graphs-1","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"","category":"section"},{"location":"man/kroneckerpowers/#Kronecker-powers-1","page":"Kronecker powers and graphs","title":"Kronecker powers","text":"","category":"section"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"Repeated Kronecker multiplications of the same matrix, i.e.","category":"page"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"A^otimes n=otimes_i=1^n A = underbraceA otimes A otimes ldots otimes A_ntext times","category":"page"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"Kronecker powers are supported using kronecker(A, n) or, equivalently, ⊗(A, n). These functions yield an instance of KroneckerPower, as struct which holds the matrix and the power. It works just like instances of KroneckerProduct, but more efficient since only a single matrix has to be stored and manipulated. These products work as expected.","category":"page"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"using Kronecker\nA = rand(2, 2)\nK2 = A ⊗ 2\ninv(K2)\nK12 = K2 ⊗ 6  # works recursively\nK12^2  # example","category":"page"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"kronecker(A::AbstractMatrix, pow::Int)\n⊗(A::AbstractMatrix, pow::Int)","category":"page"},{"location":"man/kroneckerpowers/#Kronecker.kronecker-Tuple{AbstractArray{T,2} where T,Int64}","page":"Kronecker powers and graphs","title":"Kronecker.kronecker","text":"kronecker(A::AbstractMatrix, pow::Int)\n\nKronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.\n\n\n\n\n\n","category":"method"},{"location":"man/kroneckerpowers/#Kronecker.:⊗-Tuple{AbstractArray{T,2} where T,Int64}","page":"Kronecker powers and graphs","title":"Kronecker.:⊗","text":"⊗(A::AbstractMatrix, pow::Int)\n\nKronecker power, computes A ⊗ A ⊗ ... ⊗ A. Returns a lazy KroneckerPower type.\n\n\n\n\n\n","category":"method"},{"location":"man/kroneckerpowers/#Kronecker-graphs-1","page":"Kronecker powers and graphs","title":"Kronecker graphs","text":"","category":"section"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"An exciting application of Kronecker powers (or Krocker products in general) is generating large, realistic graphs from an initial 'seed' via a stochastic process. This is called a Kronecker graph and is described in detail in Leskovec et al. (2008).","category":"page"},{"location":"man/kroneckerpowers/#","page":"Kronecker powers and graphs","title":"Kronecker powers and graphs","text":"If we use an initial matrix P with values in 01 as a seed, then P_n=P^otimes n can be seen as a probability distribution for an adjacency matrix of a graph. The elements (P_n)_ij give the probabilities that there is an edge from node i to node j. Leskovec and co-authors give two algori
[…]
