### 2019-11-12 3c52c4d
:put_litter_in_its_place: remove documenter dependency
