### 2019-11-12 e25f982
better symbol for joined names
