### 2019-11-15 0ccae15
added efficient backslash solve
