+++ src/base.jl
@@ -307,3 +307,13 @@ function Base.:*(K::AbstractKroneckerProduct, a::Number)
     A, B = getmatrices(K)
     kronecker(A, B * a)
 end
+
+# SOLVING
+using LinearAlgebra: checksquare
+function LinearAlgebra.:\(K::AbstractKroneckerProduct{T}, c::AbstractVector{T}) where {T}
+    !issquare(K.A) && throw(DimensionMismatch("matrix A is not square: dimensions are " * size(K.A)))
+    !issquare(K.B) && throw(DimensionMismatch("matrix B is not square: dimensions are " * size(K.B)))
+    size(K, 2) != length(c) && throw(DimensionMismatch("size(K, 2) != length(c)"))
+    C = reshape(c, size(K.B, 1), size(K.A, 1)) # matricify
+    return vec((K.B \ C) / K.A') #(A ⊗ B)vec(X) = vec(C) <=> BXA' = C => X = B^{-1} C A'^{-1}
+end
