### 2019-11-15 9cbed6c
added fallback for non-square A, B
