### 2019-11-15 751fd99
enable efficient backslash for least-squares problems
