### 2019-11-16 668f1c6
added more test cases for non-square Kronecker systems
