+++ src/base.jl
@@ -309,9 +309,8 @@ function Base.:*(K::AbstractKroneckerProduct, a::Number)
 end
 
 # SOLVING
-using LinearAlgebra: checksquare
 function LinearAlgebra.:\(K::AbstractKroneckerProduct{T}, c::AbstractVector{T}) where {T}
-    size(K, 2) != length(c) && throw(DimensionMismatch("size(K, 2) != length(c)"))
+    size(K, 1) != length(c) && throw(DimensionMismatch("size(K, 1) != length(c)"))
     C = reshape(c, size(K.B, 1), size(K.A, 1)) # matricify
     return vec((K.B \ C) / K.A') #(A ⊗ B)vec(X) = vec(C) <=> BXA' = C => X = B^{-1} C A'^{-1}
 end
