### 2019-12-07 87ae96a
remove Documenter dep
