### 2020-05-07 be59638
Remove sparse from collect and added kron for KroneckerSum
Collected kronecker sums are no longer sparse by default.
kron is added to ensure that higher order sums preserve the array types.
