### 2020-05-07 a701d1b
Test dense and sparse arrays for KroneckerSum
Test more types and variations of higher order sums
