+++ test/testkroneckersum.jl
@@ -1,77 +1,87 @@
 @testset "Kronecker sums" begin
 
+    As = (rand(4,4), sprand(4,4,1.0))
+    Bs = (rand(3,3), sprand(3,3,1.0))
+    Cs = (rand(5,5), sprand(5,5,1.0))
+    Ds = (rand(ComplexF64,6,6), sprand(ComplexF64,6,6,1.0))
+    arraytypes = (Matrix, SparseMatrixCSC)
+    for (A, B, C, D, arraytype) in zip(As, Bs, Cs, Ds, arraytypes)
+        IA = oneunit(A)
+        IB = oneunit(B)
 
-    A = rand(4,4); IA = oneunit(A)
-    B = rand(3,3); IB = oneunit(B)
+        KS = A ⊕ B
+        kronsum = kron(A, IB) + kron(IA, B)
 
-    KS = A ⊕ B
-    kronsum = kron(A, IB) + kron(IA, B)
+        @test eltype(KS) <: Float64
+        @test KS isa AbstractMatrix{Float64}
+        @test KS isa GeneralizedKroneckerProduct{Float64}
+        @test !isa(KS, AbstractKroneckerProduct)
 
-    @test eltype(KS) <: Float64
-    @test KS isa AbstractMatrix{Float64}
-    @test KS isa GeneralizedKroneckerProduct{Float64}
-    @test !isa(KS, AbstractKroneckerProduct)
+        @test collect(KS) ≈ kronsum
+        @test collect(KS) isa arraytype
 
-    @test collect(KS) ≈ kronsum
-    @test collect(KS) isa SparseMatrixCSC
+        @test issquare(KS)
 
-    @test issquare(KS)
+        IC = oneunit(C)
+        KS3 = A ⊕ B ⊕ C
+        KS3AB = (A ⊕ B) ⊕ C 
+        KS3BC = A ⊕ (B ⊕ C)
+        
+        kronsum3 = kron(A,IB,IC) + kron(IA,B,IC) + kron(IA,IB,C)
 
-    C = rand(5,5); IC = oneunit(C)
-    KS3 = A ⊕ B ⊕ C
-    kronsum3 = kron(A,IB,IC) + kron(IA,B,IC) + kron(IA,IB,C)
+        for ks3 in [KS3, KS3AB, KS3BC]
+            @test collect(ks3) ≈ kronsum3
+            @test collect(ks3) isa arraytype
+            @test order(ks3) == 3
+            @test getindex(ks3,2,3) == kronsum3[2,3]
+        end
+        
+        @test order(KS) == 2
+        @test getmatrices(KS) == (A,B)
+        @test getindex(KS,2,3) == kronsum[2,3]
+        
+        
+        ID = oneunit(D)
 
-    @test collect(KS3) ≈ kronsum3
-    @test collect(KS3) isa SparseMatrixCSC
-
-    @test order(KS) == 2
-    @test order(KS3) == 3
-
-    @test getmatrices(KS) == (A,B)
-
-    @test getindex(KS,2,3) == kronsum[2,3]
-    @test getindex(KS3,2,3) == kronsum3[2,3]
-
-    D = rand(ComplexF64,6,6); ID = oneunit(D)
-
-    @testset "Structure of sums" begin
-        @test size(A ⊕ D) == size(A) .* size(D,1)
-        @test size(B ⊕ C ⊕ D) == size(B) .* size(C) .* size(D)
-        @test eltype(A ⊕ B) == Float64
-        @test eltype(C ⊕ D) == ComplexF64
-    end
-
-    @testset "Basic linear algebra for sums" begin
-        @test tr(KS) ≈ tr(kronsum)
-        @test KS' == kronsum'
[…]
