### 2020-05-07 2bb363b
Fixed deprecated 'names' to 'dimnames'
