### 2020-05-07 71bb487
:pencil2: Cleaning up collect function and docstring
