+++ src/kroneckersum.jl
@@ -108,11 +108,10 @@ end
     collect(K::AbstractKroneckerSum)
 
 Collects a lazy instance of the `AbstractKroneckerSum` type into a full,
-native matrix. Returns the result as a sparse matrix.
+native matrix.
 """
 function Base.collect(K::AbstractKroneckerSum)
     A, B = getmatrices(K)
-    #A, B = sparse(A), sparse(B)
     IA, IB = oneunit(A), oneunit(B)
     return kron(A, IB) + kron(IA, B)
 end
