### 2020-05-10 28b4eb5
:art: restructured kron containing KroneckerSums
