### 2020-05-10 9493299
:sparkles: sparse utility function for KroneckerSum
