### 2020-05-10 0262919
:white_check_mark: Add test when combining kron with KroneckerSum
