### 2020-05-10 a40d43c
:zap: Collect K::KroneckerSum with getmatrices(sparse(K))
Avoids calling `sparse(K.A), sparse(K.B)` when `K.A` and `K.B` are 
already sparse
