+++ src/kroneckersum.jl
@@ -111,8 +111,7 @@ Collects a lazy instance of the `AbstractKroneckerSum` type into a full,
 native matrix. Returns the result as a sparse matrix.
 """
 function Base.collect(K::AbstractKroneckerSum)
-    A, B = getmatrices(K)
-    #A, B = sparse(A), sparse(B)
+    A, B = getmatrices(sparse(K))
     IA, IB = oneunit(A), oneunit(B)
     return kron(A, IB) + kron(IA, B)
 end
