### 2020-05-10 63a1442
:rewind: Revert ":pencil2: Cleaning up collect function and docstring"
This reverts commit 71bb487b33babcf723f8ce3e80a9748e1b80b49c.
