### 2020-05-10 f285086
:white_check_mark: Fix type test for collected KroneckerSum
