### 2020-05-10 307ac9c
:white_check_mark: Fix type test (higher order collect for KroneckerSum)
