### 2020-05-16 585711e
:tada: fallback of `collect!`
