### 2020-05-16 b0b7103
:wrench: unit tests `collect!`
