### 2020-05-17 51fe00b
:dragon: workingish version of `collect!` kronecker powers
