+++ test/testkroneckersum.jl
@@ -5,7 +5,7 @@
     B = rand(3,3); IB = oneunit(B)
 
     KS = A ⊕ B
-    kronsum = kron(A, IB) + kron(IA, B)
+    kronsum = kron(A, IB) .+ kron(IA, B)
 
     @test eltype(KS) <: Float64
     @test KS isa AbstractMatrix{Float64}
@@ -13,7 +13,10 @@
     @test !isa(KS, AbstractKroneckerProduct)
 
     @test collect(KS) ≈ kronsum
+    @test collect!(similar(kronsum), KS) ≈ kronsum
     @test collect(KS) isa SparseMatrixCSC
+
+
 
     @test issquare(KS)
 
