### 2020-05-18 3ec839d
:ok_hand: Remove type annotations
