### 2020-05-18 c41360d
:ok_hand: :pencil: Add docstring for issparse
