+++ src/base.jl
@@ -26,7 +26,8 @@ end
     collect(K::GeneralizedKroneckerProduct)
 
 Collects a lazy instance of the `GeneralizedKroneckerProduct` type into a dense,
-native matrix. General element-wise fallback.
+native matrix. Falls back to the element-wise case when not specialized method
+is defined.
 """
 function collect(K::GeneralizedKroneckerProduct{T}) where {T}
     C = Matrix{T}(undef, size(K)...)
@@ -256,6 +257,7 @@ end
 
 # COLLECTING
 
+#=
 """
     collect(K::AbstractKroneckerProduct)
 
@@ -266,8 +268,11 @@ function collect(K::AbstractKroneckerProduct)
     A, B = getmatrices(K)
     return kron(A, B)
 end
+=#
 
-function mykron!(C::AbstractArray, A::AbstractArray, B::AbstractArray)
+
+# function for in-place Kronecker product
+function _kron!(C::AbstractArray, A::AbstractArray, B::AbstractArray)
     m = 0
     @inbounds for j = 1:size(A,2), l = 1:size(B,2), i = 1:size(A,1)
         Aij = A[i,j]
@@ -278,38 +283,22 @@ function mykron!(C::AbstractArray, A::AbstractArray, B::AbstractArray)
     return C
 end
 
+_kron!(C::AbstractArray, A::GeneralizedKroneckerProduct, B::AbstractArray) = _kron!(C, collect(A), B)
+_kron!(C::AbstractArray, A::AbstractArray, B::GeneralizedKroneckerProduct) = _kron!(C, A, collect(B))
+_kron!(C::AbstractArray, A::GeneralizedKroneckerProduct, B::GeneralizedKroneckerProduct) = _kron!(C, collect(A), collect(B))
+
+"""
+    collect!(C::AbstractMatrix, K::AbstractKroneckerProduct)
+
+In-place collection of `K` in `C` where `K` is an `AbstractKroneckerProduct`, i.e.,
+`K = A ⊗ B`.
+"""
 function collect!(C::AbstractMatrix, K::AbstractKroneckerProduct)
     size(C) == size(K) || throw(DimensionMismatch("`K` $(size(K)) cannot be collected in `C` $(size(C))"))
     A, B = getmatrices(K)
-    return mykron!(C, A, B)
+    return _kron!(C, A, B)
 end
 
-
-
-
-function tile!(C::AbstractMatrix, A::AbstractMatrix, r=1, s=1)
-    n, m = size(A)
-    N, M = size(C)
-    k = 0
-    @inbounds for q in 1:s:div(M, m), j in 1:m, repc in 1:s
-        @inbounds for p in 1:r:div(N, n), i in 1:n, repr in 1:r
-            C[k+=1] = A[i,j]
-        end
-    end
-    return C
-end
-
-function tileprod!(C::AbstractMatrix, A::AbstractMatrix, r=1, s=1)
-    n, m = size(A)
-    N, M = size(C)
-    k = 0
-    @inbounds for q in 1:s:div(M, m), j in 1:m, repc in 1:s
-        @inbounds for p in 1:r:div(N, n), i in 1:n, repr in 1:r
-            C[k+=1] *= A[i,j]
-        end
-    end
-    return C
-end
 
 """
     Matrix(K::GeneralizedKroneckerProduct)
