### 2020-07-06 bc6e874
:zap: `collect!` for KroneckerSum
