+++ test/testkroneckersum.jl
@@ -1,10 +1,12 @@
 @testset "Kronecker sums" begin
+    using SparseArrays
 
     As = (rand(4,4), sprand(4,4,1.0))
     Bs = (rand(3,3), sprand(3,3,1.0))
     Cs = (rand(5,5), sprand(5,5,1.0))
     Ds = (rand(ComplexF64,6,6), sprand(ComplexF64,6,6,1.0))
     arraytypes = (Matrix, SparseMatrixCSC)
+
     for (A, B, C, D, arraytype) in zip(As, Bs, Cs, Ds, arraytypes)
         IA = oneunit(A)
         IB = oneunit(B)
@@ -36,42 +38,10 @@
             @test getindex(ks3,2,3) == kronsum3[2,3]
         end
 
-    A = rand(4,4); IA = oneunit(A)
-    B = rand(3,3); IB = oneunit(B)
-
-    KS = A ⊕ B
-    kronsum = kron(A, IB) .+ kron(IA, B)
-
-    @test eltype(KS) <: Float64
-    @test KS isa AbstractMatrix{Float64}
-    @test KS isa GeneralizedKroneckerProduct{Float64}
-    @test !isa(KS, AbstractKroneckerProduct)
-
-    @test collect(KS) ≈ kronsum
-    @test collect!(similar(kronsum), KS) ≈ kronsum
-    @test collect(KS) isa SparseMatrixCSC
-
-    @test issquare(KS)
-
-    C = rand(5,5); IC = oneunit(C)
-    KS3 = A ⊕ B ⊕ C
-    kronsum3 = kron(A,IB,IC) + kron(IA,B,IC) + kron(IA,IB,C)
-
-    @test collect(KS3) ≈ kronsum3
-    @test collect(KS3) isa SparseMatrixCSC
-
-    @test order(KS) == 2
-    @test order(KS3) == 3
-
-    @test getmatrices(KS) == (A,B)
-
-    @test getindex(KS,2,3) == kronsum[2,3]
-    @test getindex(KS3,2,3) == kronsum3[2,3]
-    
         @test order(KS) == 2
         @test getmatrices(KS) == (A,B)
         @test getindex(KS,2,3) == kronsum[2,3]
-        
+
         ID = oneunit(D)
 
         @testset "Structure of sums" begin
