+++ src/kroneckersum.jl
@@ -190,14 +190,6 @@ function Base.exp(K::AbstractKroneckerSum)
     return kronecker(exp(A), exp(B))
 end
 
-#=
-function Base.sum(K::KroneckerSum)
-    A, B = getmatrices(K)
-    n, m = size(A, 1), size(B, 1)
-    return m * sum(A) + n * sum(B)
-end
-=#
-
 function Base.sum(K::KroneckerSum; dims::Union{Int,Nothing}=nothing)
     A, B = getmatrices(K)
     n, m = size(A, 1), size(B, 1)
@@ -205,7 +197,7 @@ function Base.sum(K::KroneckerSum; dims::Union{Int,Nothing}=nothing)
         return kron(sum(A, dims=1), ones(1, m)) .+ kron(ones(1, n), sum(B, dims=1))
     elseif dims==2
         return kron(sum(A, dims=2), ones(m, 1)) .+ kron(ones(n, 1), sum(B, dims=2))
-    elseif dims==nothing
+    elseif dims isa Nothing
         m * sum(A) + n * sum(B)
     else
         throw(ArgumentError("`dims` should be 1 or 2"))
